support for custom controllers
