+++ src/mcp/index.ts
@@ -3,3 +3,5 @@ export * from './interfaces';
 export * from './mcp.module';
 export * from './services/mcp-registry.service';
 export * from './services/mcp-executor.service';
+export * from './services/mcp-sse.service';
+export * from './services/mcp-streamable-http.service';
